Create a comments system, add the tagging functionality
Create a comment system for the blog: build a model, create a form from the model,
handle the ModelForm in views, add comments to the post detail template.
Add the tagging functionality using Django Taggit app.
Build a functionality to display similar posts by the number of tags they share.