Add sitemap, create feed for blog posts
Add a sitemap to the site by adding the sitemap framework to the installed apps.
Create a file sitemaps.py inside the blog application that contains a custom
sitemap. Add the sitemap url to the site url file.
Create feeds for blog posts by creating a file feeds.py in the blog app that
contains a LatestPostsFeed class. Instantiate the feed in a new URL pattern.
Add a feed subscription link to the blog sidebar in the base.html file.